# deliverables/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Sentinel AML \u2014 Transaction Monitoring System",
        "description": "Real-time money-laundering detection for MeridianTrust Bank.\n\n**Flow:** ingestion \u2192 detection \u2192 risk-scored alert \u2192 case disposition.\n\n**Roles** (RBAC is enforced at this API layer, not only in the UI):\n- `ANALYST` \u2014 alert queue (PII masked), cases, dispositions\n- `SENIOR_ANALYST` \u2014 the above, plus unmasked PII and SAR escalation\n- `COMPLIANCE_ADMIN` \u2014 the above, plus ingestion and rule configuration\n\nDetection thresholds are tunable at runtime via\n`PATCH /api/v1/admin/rules/{ruleCode}` \u2014 no redeployment required.\n",
        "contact": {
            "name": "MeridianTrust Financial Crime Engineering"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current host"
        }
    ],
    "security": [
        {
            "basicAuth": []
        }
    ],
...
    "paths": {
...
        "/api/v1/cases/{caseRef}/status": {
            "patch": {
                "tags": [
                    "Cases"
                ],
                "summary": "Transition a case",
                "description": "Closing a case requires a disposition and reason, and cascades that\ndisposition onto every member alert \u2014 each cascade separately audited.\n",
                "operationId": "transition",
                "parameters": [
                    {
                        "name": "caseRef",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransitionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Transition applied",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Detail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing disposition or reason",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Detail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Role does not permit this disposition",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Detail"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Illegal transition for the current state",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Detail"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/alerts/{alertRef}/status": {
            "patch": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Transition an alert",
                "description": "Moves an alert through its lifecycle. Closing requires a disposition\nand a reason, both retained permanently with the analyst's identity \u2014\nalerts are never deleted (business rule 6).\n\nESCALATED_TO_SAR requires the SENIOR_ANALYST role.\n",
                "operationId": "transition_1",
                "parameters": [
                    {
                        "name": "alertRef",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlertTransitionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Transition applied",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetailDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing disposition or reason",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetailDto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Role does not permit this disposition",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetailDto"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Concurrently modified \u2014 reload and retry",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetailDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Illegal transition for the current state",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetailDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/alerts": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Alert queue",
                "description": "Returns alerts with customer PII masked (business rule 8).\nSorted by risk score descending by default so the highest-risk\nalerts reach the top of the queue (business rule 7).\n",
                "operationId": "queue_1",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "OPEN, IN_REVIEW, ESCALATED or CLOSED",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "OPEN",
                                "IN_REVIEW",
                                "ESCALATED",
                                "CLOSED"
                            ]
                        }
                    },
                    {
                        "name": "severity",
                        "in": "query",
                        "description": "LOW, MEDIUM, HIGH or CRITICAL",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "LOW",
                                "MEDIUM",
                                "HIGH",
                                "CRITICAL"
                            ]
                        }
                    },
                    {
                        "name": "ruleCode",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customerId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "minScore",
                        "in": "query",
                        "description": "Only alerts at or above this risk score",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 25
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Property to sort by, e.g. riskScore or lastDetectedAt",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "riskScore"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseAlertSummaryDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/alerts/{alertRef}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Alert detail",
                "description": "Full explanation, supporting evidence transactions and risk score breakdown.",
                "operationId": "detail_1",
                "parameters": [
                    {
                        "name": "alertRef",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetailDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No such alert",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetailDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AlertSummaryDto": {
                "type": "object",
                "description": "Alert queue row. Customer PII is masked; use the detail endpoint for full values.",
                "properties": {
                    "alertRef": {
                        "type": "string",
                        "example": "ALT-20260919-A1B2C3D4"
                    },
                    "customerId": {
                        "type": "string",
                        "example": "CUST_00042"
                    },
                    "customerNameMasked": {
                        "type": "string",
                        "description": "Masked customer name",
                        "example": "K****** S*****"
                    },
                    "accountMasked": {
                        "type": "string",
                        "description": "Masked account number",
                        "example": "******0123"
                    },
                    "ruleCode": {
                        "type": "string"
                    },
                    "typology": {
                        "type": "string"
                    },
                    "riskScore": {
                        "type": "integer",
                        "format": "int32",
                        "description": "0-100 weighted risk score (business rule 7)",
                        "example": 78
                    },
                    "severity": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "evidenceAmountBase": {
                        "type": "number"
                    },
                    "evidenceCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "triggerCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "disposition": {
                        "type": "string"
                    },
                    "firstDetectedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastDetectedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "AuditEntry": {
                "type": "object",
                "description": "One immutable audit record.",
                "properties": {
                    "action": {
                        "type": "string"
                    },
                    "actor": {
                        "type": "string"
                    },
                    "fromState": {
                        "type": "string"
                    },
                    "toState": {
                        "type": "string"
                    },
                    "details": {
                        "type": "string"
                    },
                    "occurredAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Detail": {
                "type": "object",
                "description": "Full case view including member alerts and audit trail.",
                "properties": {
                    "caseRef": {
                        "type": "string"
                    },
                    "customerId": {
                        "type": "string"
                    },
                    "customerNameMasked": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "priority": {
                        "type": "string"
                    },
                    "aggregateRiskScore": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "assignedTo": {
                        "type": "string"
                    },
                    "narrative": {
                        "type": "string"
                    },
                    "disposition": {
                        "type": "string"
                    },
                    "dispositionReason": {
                        "type": "string"
                    },
                    "openedBy": {
                        "type": "string"
                    },
                    "openedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "closedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "alerts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AlertSummaryDto"
                        }
                    },
                    "auditTrail": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AuditEntry"
                        }
                    }
                }
            },
...
            "TransitionRequest": {
                "type": "object",
                "description": "Case state transition. Closing requires a disposition and reason.",
                "properties": {
                    "targetStatus": {
                        "type": "string",
                        "enum": [
                            "NEW",
                            "INVESTIGATING",
                            "PENDING_APPROVAL",
                            "CLOSED"
                        ]
                    },
                    "disposition": {
                        "type": "string",
                        "enum": [
                            "TRUE_POSITIVE",
                            "FALSE_POSITIVE",
                            "ESCALATED_TO_SAR",
                            "CLEARED_NO_ACTION"
                        ]
                    },
                    "reason": {
                        "type": "string",
                        "maxLength": 4000,
                        "minLength": 0
                    }
                },
                "required": [
                    "targetStatus"
                ]
            },
...
            "AlertTransitionRequest": {
                "type": "object",
                "description": "Alert status transition. Closing requires a disposition and reason (business rule 6).",
                "properties": {
                    "targetStatus": {
                        "type": "string",
                        "enum": [
                            "OPEN",
                            "IN_REVIEW",
                            "ESCALATED",
                            "CLOSED"
                        ]
                    },
                    "disposition": {
                        "type": "string",
                        "description": "Required when targetStatus is CLOSED",
                        "enum": [
                            "TRUE_POSITIVE",
                            "FALSE_POSITIVE",
                            "ESCALATED_TO_SAR",
                            "CLEARED_NO_ACTION"
                        ]
                    },
                    "reason": {
                        "type": "string",
                        "description": "Required when targetStatus is CLOSED",
                        "example": "Verified as salary payments",
                        "maxLength": 2000,
                        "minLength": 0
                    }
                },
                "required": [
                    "targetStatus"
                ]
            },
            "AlertDetailDto": {
                "type": "object",
                "description": "Alert detail with evidence transactions and risk score breakdown.",
                "properties": {
                    "alertRef": {
                        "type": "string"
                    },
                    "customerId": {
                        "type": "string"
                    },
                    "customerName": {
                        "type": "string"
                    },
                    "accountId": {
                        "type": "string"
                    },
                    "ruleCode": {
                        "type": "string"
                    },
                    "typology": {
                        "type": "string"
                    },
                    "riskScore": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "severity": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "explanation": {
                        "type": "string",
                        "description": "Human-readable justification generated from the evidence"
                    },
                    "scoreBreakdown": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "description": "How the risk score was composed"
                    },
                    "evidence": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EvidenceTransactionDto"
                        }
                    },
                    "evidenceAmountBase": {
                        "type": "number"
                    },
                    "triggerCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "disposition": {
                        "type": "string"
                    },
                    "dispositionReason": {
                        "type": "string"
                    },
                    "disposedBy": {
                        "type": "string",
                        "description": "Analyst who disposed the alert (business rule 6)"
                    },
                    "disposedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "caseRef": {
                        "type": "string"
                    },
                    "firstDetectedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastDetectedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "EvidenceTransactionDto": {
                "type": "object",
                "properties": {
                    "transactionId": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "direction": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "amountBase": {
                        "type": "number"
                    },
                    "channel": {
                        "type": "string"
                    },
                    "counterpartyName": {
                        "type": "string"
                    },
                    "counterpartyCountry": {
                        "type": "string"
                    }
                }
            },
...
            "PageResponseAlertSummaryDto": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AlertSummaryDto"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            }
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "description": "HTTP Basic. Demo users: analyst / senior / admin.",
                "scheme": "basic"
            }
        }
    }
}